Enhance agent functionality by adding RESPONSE_TAG to agent exports, updating AzureChatOpenAI method to accept tags, and refactoring load_prompt_files into cached_loader. Remove unused imports and streamline message handling in SingleAgent.
