+++ src/agents/__init__.py
@@ -4,4 +4,6 @@ from .agent import Agent
 from .agent_registry import agent_registry
 from .single_agent import SingleAgent
 
-__all__ = ["agent_registry", "Agent", "SingleAgent"]
+RESPONSE_TAG = "response"
+
+__all__ = ["agent_registry", "Agent", "SingleAgent", "RESPONSE_TAG"]
