+++ src/agents/graph_agent.py
@@ -19,7 +19,6 @@ from langgraph.graph.message import add_messages
 from streamlit.logger import get_logger
 from typing_extensions import Annotated
 
-from utils.openai_utils import load_prompt_files
 from utils.streamlit_context import with_streamlit_context
 
 from .agent import Agent
